+++ assets/src/scripts/components/hydrograph/cursor.js
@@ -109,6 +109,7 @@ export const cursorSlider = function (elem) {
             wrap.append('input')
                 .attr('type', 'range')
                 .attr('id', 'cursor-slider')
+                .attr('class', 'usa-range')
                 .attr('aria-label', 'Hydrograph cursor slider')
                 .on('input', dispatch(function () {
                     return Actions.setCursorOffset(this.valueAsNumber);