+++ assets/src/scripts/components/hydrograph/parameters.js
@@ -140,14 +140,17 @@ export const plotSeriesSelectTable = function (elem, {siteno, availableTimeSerie
     const columnHeaders = ['Parameter', 'Preview', '#'];
     const tableContainer = elem.append('div')
         .attr('id', 'select-time-series');
+
     tableContainer.append('label')
         .attr('id', 'select-time-series-label')
         .text('Select a time series');
     const table = tableContainer.append('table')
-        .classed('usa-table-borderless', true)
+        .classed('usa-table', true)
+        .classed('usa-table--borderless', true)
         .attr('aria-labelledby', 'select-time-series-label')
         .attr('tabindex', 0)
         .attr('role', 'listbox');
+
 
     table.append('thead')
         .append('tr')