+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add Docker config for server-rendering PNGs
 - Use a Puppeteer/Chrome process pool for server rendering
 - Map now contains Leaflet layer control to toggle satellite and gray with hydro layers
+- Project uses wdfn-viz npm package to provide USGS visual identification. This required updating 
+the project styles to use USWDS 2.0.x
 
 ### Changed
 - Links in monitoring-location now use configurable endpoints.